+++ tests/Fixtures/Models/Author.php
@@ -6,6 +6,8 @@ use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
 use Illuminate\Database\Eloquent\Relations\BelongsToMany;
 use Illuminate\Database\Eloquent\Relations\HasMany;
+use Illuminate\Database\Eloquent\Relations\HasOne;
+use Illuminate\Database\Eloquent\Relations\MorphMany;
 use NormCache\Traits\Cacheable;
 
 class Author extends Model
@@ -27,5 +29,15 @@ class Author extends Model
     public function tags(): BelongsToMany
     {
         return $this->belongsToMany(Tag::class, 'author_tag');
+    }
+
+    public function firstPost(): HasOne
+    {
+        return $this->hasOne(Post::class);
+    }
+
+    public function comments(): MorphMany
+    {
+        return $this->morphMany(Comment::class, 'commentable');
     }
 }